requirements.txt
pymupdf==1.24.0
pytesseract==0.3.10
Pillow==10.3.0
rapidfuzz==3.9.0
pydantic==2.7.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.7.0
pandas>=2.0.0
